refactor: renamed dialog-host files to dialog-message-content
