+++ libs/portal-integration-angular/src/index.ts
@@ -51,7 +51,7 @@ export * from './lib/core/components/data-list-grid-sorting/data-list-grid-sorti
 export * from './lib/core/components/group-by-count-diagram/group-by-count-diagram.component'
 export * from './lib/core/components/diagram/diagram.component'
 export * from './lib/core/components/button-dialog/button-dialog.component'
-export * from './lib/core/components/button-dialog/dialog-host/dialog-host.component'
+export * from './lib/core/components/button-dialog/dialog-message-content/dialog-message-content.component'
 export * from './lib/core/components/search-config/search-config.component'
 export * from './lib/core/components/loading-indicator/loading-indicator.component'
 export * from './lib/core/components/ocx-content-container/ocx-content-container.component'
