+++ libs/portal-integration-angular/src/lib/core/components/button-dialog/dialog-message-content/dialog-message-content.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from '@angular/core'
 
 @Component({
-  templateUrl: `./dialog-host.component.html`,
+  templateUrl: `./dialog-message-content.component.html`,
 })
 export class DialogMessageContentComponent {
   @Input() message = 'message'
